requirements/testing.txt
pytest-cov
pytest-django
pytest-watch
